Use read_bytes() (not _io:read()) for better error messages
self._io:read() does not throw an error if the number of requested bytes
is not available; in this case, it simply returns fewer bytes than
requested. This is not good because it will still fail, only at a higher
level - e.g. in unpack() for primitive reads:

```
../runtime/lua/kaitaistruct.lua:145: bad argument #2 to 'unpack' (data string too short)
```

This is a cryptic error (it looks like an internal error as if the
runtime library broke, which basically happened - it wasn't prepared for
this), completely dependent on the current implementation and
inconsistent among other manifestations of the EOF condition. This makes
it hard to catch all EOF errors regardless of the function in which they
occur - to allow that, EOF errors from all methods must behave the same.

Unlike the low-level self._io:read(), self:read_bytes() does check that
all required bytes were received and throws a uniform error if not.

The same pattern can be seen in the Python runtime lib (see
https://github.com/kaitai-io/kaitai_struct_python_runtime/blob/924f3251/kaitaistruct.py#L182)
- read_bytes() is used everywhere (not _io.read()) because of the EOF
check it provides.